Update install.sh
This should fix the install script not running on OSMC on the Pi 3 and 2. Testing if it's running on ARM 6 or 7 rather than testing if it's running on Rapbian, as OSMC for Pi 2/3 identifies itself as Debian not Raspbian. This could cause the script to run on other ARM devices I suppose, but it's better than failing on some Pis. If anyone knows a better way to do this so that it only detects Pis then feel free to change it.